moved get_logger functionality into config_metplus so it doesn't need to be called explicitly after config setup
