+++ metplus/util/met_util.py
@@ -92,7 +92,7 @@ def pre_run_setup(config_inputs):
     # Read config inputs and return a config instance
     config = config_metplus.setup(config_inputs)
 
-    logger = get_logger(config)
+    logger = config.logger
 
     config.set('config', 'METPLUS_VERSION', version_number)
     logger.info('Running METplus v%s called with command: %s',
@@ -1122,194 +1122,6 @@ def rmtree(tree, logger=None):
     if logger:
         logger.info('%s: rmtree' % (tree,))
     shutil.rmtree(tree, ignore_errors=False)
-
-
-def set_logvars(config, logger=None):
-    """!Sets and adds the LOG_METPLUS and LOG_TIMESTAMP
-       to the config object. If LOG_METPLUS was already defined by the
-       user in their conf file. It expands and rewrites it in the conf
-       object and the final file.
-       conf file.
-       Args:
-           @param config:   the config instance
-           @param logger: the logger, optional
-    """
-
-    if logger is None:
-        logger = config.log()
-
-    # LOG_TIMESTAMP_TEMPLATE is not required in the conf file,
-    # so lets first test for that.
-    log_timestamp_template = config.getstr('config', 'LOG_TIMESTAMP_TEMPLATE', '')
-    if log_timestamp_template:
-        # Note: strftime appears to handle if log_timestamp_template
-        # is a string ie. 'blah' and not a valid set of % directives %Y%m%d,
-        # it does return the string 'blah', instead of crashing.
-        # However, I'm still going to test for a valid % directive and
-        # set a default. It probably is ok to remove the if not block pattern
-        # test, and not set a default, especially if causing some unintended
-        # consequences or the pattern is not capturing a valid directive.
-        # The reality is, the user is expected to have entered a correct
-        # directive in the conf file.
-        # This pattern is meant to test for a repeating set of
-        # case insensitive %(AnyAlphabeticCharacter), ie. %Y%m ...
-        # The basic pattern is (%+[a-z])+ , %+ allows for 1 or more
-        # % characters, ie. %%Y, %% is a valid directive.
-        # (?i) case insensitive, \A begin string \Z end of string
-        if not re.match(r'(?i)\A(?:(%+[a-z])+)\Z', log_timestamp_template):
-            logger.warning('Your LOG_TIMESTAMP_TEMPLATE is not '
-                           'a valid strftime directive: %s' % repr(log_timestamp_template))
-            logger.info('Using the following default: %Y%m%d%H')
-            log_timestamp_template = '%Y%m%d%H'
-        date_t = datetime.datetime.now()
-        if config.getbool('config', 'LOG_TIMESTAMP_USE_DATATIME', False):
-            if is_loop_by_init(config):
-                date_t = datetime.datetime.strptime(config.getstr('config',
-                                                                  'INIT_BEG'),
-                                                    config.getstr('config',
-                                                                  'INIT_TIME_FMT'))
-            else:
-                date_t = datetime.datetime.strptime(config.getstr('config',
-                                                                  'VALID_BEG'),
-                                                    config.getstr('config',
-                                                                  'VALID_TIME_FMT'))
-        log_filenametimestamp = date_t.strftime(log_timestamp_template)
-    else:
-        log_filenametimestamp = ''
-
-    log_dir = config.getdir('LOG_DIR')
-
-    # NOTE: LOG_METPLUS or metpluslog is meant to include the absolute path
-    #       and the metpluslog_filename,
-    # so metpluslog = /path/to/metpluslog_filename
-
-    # if LOG_METPLUS =  unset in the conf file, means NO logging.
-    # Also, assUmes the user has included the intended path in LOG_METPLUS.
-    user_defined_log_file = None
-    if config.has_option('config', 'LOG_METPLUS'):
-        user_defined_log_file = True
-        # strinterp will set metpluslog to '' if LOG_METPLUS =  is unset.
-        metpluslog = config.strinterp('config', '{LOG_METPLUS}',
-                                      LOG_TIMESTAMP_TEMPLATE=log_filenametimestamp)
-
-        # test if there is any path information, if there is, assUme it is as intended,
-        # if there is not, than add log_dir.
-        if metpluslog:
-            if os.path.basename(metpluslog) == metpluslog:
-                metpluslog = os.path.join(log_dir, metpluslog)
-    else:
-        # No LOG_METPLUS in conf file, so let the code try to set it,
-        # if the user defined the variable LOG_FILENAME_TEMPLATE.
-        # LOG_FILENAME_TEMPLATE is an 'unpublished' variable - no one knows
-        # about it unless you are reading this. Why does this exist ?
-        # It was from my first cycle implementation. I did not want to pull
-        # it out, in case the group wanted a stand alone metplus log filename
-        # template variable.
-
-        # If metpluslog_filename includes a path, python joins it intelligently.
-        # Set the metplus log filename.
-        # strinterp will set metpluslog_filename to '' if LOG_FILENAME_TEMPLATE =
-        if config.has_option('config', 'LOG_FILENAME_TEMPLATE'):
-            metpluslog_filename = config.strinterp('config', '{LOG_FILENAME_TEMPLATE}',
-                                                   LOG_TIMESTAMP_TEMPLATE=log_filenametimestamp)
-        else:
-            metpluslog_filename = ''
-        if metpluslog_filename:
-            metpluslog = os.path.join(log_dir, metpluslog_filename)
-        else:
-            metpluslog = ''
-
-
-
-    # Adding LOG_TIMESTAMP to the final configuration file.
-    logger.info('Adding: config.LOG_TIMESTAMP=%s' % repr(log_filenametimestamp))
-    config.set('config', 'LOG_TIMESTAMP', log_filenametimestamp)
-
-    # Setting LOG_METPLUS in the configuration object
-    # At this point LOG_METPLUS will have a value or '' the empty string.
-    if user_defined_log_file:
-        logger.info('Replace [config] LOG_METPLUS with %s' % repr(metpluslog))
-    else:
-        logger.info('Adding: config.LOG_METPLUS=%s' % repr(metpluslog))
-    # expand LOG_METPLUS to ensure it is available
-    config.set('config', 'LOG_METPLUS', metpluslog)
-
-
-def get_logger(config, sublog=None):
-    """!This function will return a logger with a formatted file handler
-    for writing to the LOG_METPLUS and it sets the LOG_LEVEL. If LOG_METPLUS is
-    not defined, a logger is still returned without adding a file handler,
-    but still setting the LOG_LEVEL.
-       Args:
-           @param config:   the config instance
-           @param sublog the logging subdomain, or None
-       Returns:
-           logger: the logger
-    """
-    from . import config_metplus
-    # Retrieve all logging related parameters from the param file
-    log_dir = config.getdir('LOG_DIR')
-    log_level = config.getstr('config', 'LOG_LEVEL')
-
-    # TODO review, use builtin produtil.fileop vs. mkdir_p ?
-    # import produtil.fileop
-    # produtil.fileop.makedirs(log_dir,logger=None)
-
-    # Check if the directory path for the log file exists, if
-    # not create it.
-    if not os.path.exists(log_dir):
-        mkdir_p(log_dir)
-
-    if sublog is not None:
-        logger = config.log(sublog)
-    else:
-        logger = config.log()
-
-    # Setting of the logger level from the config instance.
-    # Check for log_level by Integer or LevelName.
-    # Try to convert the string log_level to an integer and use that, if
-    # it can't convert then we assume it is a valid LevelName, which
-    # is what is should be anyway,  ie. DEBUG.
-    # Note:
-    # Earlier versions of python2 require setLevel(<int>), argument
-    # to be an int. Passing in the LevelName, 'DEBUG' will disable
-    # logging output. Later versions of python2 will accept 'DEBUG',
-    # not sure which version that changed with, but the logic below
-    # should work for all version. I know python 2.6.6 must be an int,
-    # and python 2.7.5 accepts the LevelName.
-    try:
-        int_log_level = int(log_level)
-        logger.setLevel(int_log_level)
-    except ValueError:
-        logger.setLevel(logging.getLevelName(log_level))
-
-    # Make sure the LOG_METPLUS is defined. In this function,
-    # LOG_METPLUS should already be defined in the config object,
-    # even if it is empty, LOG_METPLUS =.
-    if not config.has_option('config', 'LOG_METPLUS'):
-        set_logvars(config)
-    metpluslog = config.getstr('config', 'LOG_METPLUS', '')
-
-    if metpluslog:
-        # It is possible that more path, other than just LOG_DIR, was added
-        # to the metpluslog, by either a user defining more path in
-        # LOG_METPLUS or LOG_FILENAME_TEMPLATE definitions in their conf file.
-        # So lets check and make more directory if needed.
-        dir_name = os.path.dirname(metpluslog)
-        if not os.path.exists(dir_name):
-            mkdir_p(dir_name)
-
-        # set up the filehandler and the formatter, etc.
-        # The default matches the oformat log.py formatter of produtil
-        # So terminal output will now match log files.
-        formatter = config_metplus.METplusLogFormatter(config)
-        file_handler = logging.FileHandler(metpluslog, mode='a')
-        file_handler.setFormatter(formatter)
-        logger.addHandler(file_handler)
-
-    # set add the logger to the config
-    config.logger = logger
-    return logger
 
 def file_exists(filename):
     """! Determines if a file exists
